[blobfs] Remove old pager ifdefs.
Now that the new pager has been enabled for some time, we can remove the
old pager code. This patch removes the ifdefs and temporary names
associated with the transition. It does not remove any of the old pager
machinery in blobfs/pager which will be refactored in a followup.

Bug: 51111
Change-Id: Ie0ee52ca97886254ed22f8121c4ec65cec3a4ecb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/560964
Reviewed-by: Brandon Castellano <bcastell@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>