[netstack] Generalize DHCP integration tests
The utility method currently used by the DHCP integration tests couples
concepts in the netemul realm with concepts in the DHCP realm. This
coupling significantly limits the topologies we can test.

This change replaces that util with one which accepts netemul
configuration and netstack/dhcp configuration independently. This allows
the callers to mix-and-match network and DHCP topologies as desired. A
side effect of this refactor is that the test body references data
allocated on the stack within the setup code. In order to support this
pattern, we moved away from the `test_case` macro (which destroys stack
allocated variables before invoking the test body).

We also removed a multi-network variant after concluding that it was
just testing non-interference across broadcast domains and therefore
didn't add anything useful for DHCP. Future work will extend the usage
of the new util to more exotic configurations.

Multiply: netstack_dhcp_integration_test
Bug: 73252
Change-Id: Ia2bb8f3d98964bc83bd3f950e15a8ee78cb906ad
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/520300
Commit-Queue: Nick Brown <nickbrow@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Amanda Tait <atait@google.com>