[build][golden_go_proto] Ignore compiler/plugin versions.
This ensures that the compiler and plugin versions, embedded inside
the generated .pb.go files, are ignored when comparing them with
the golden/reference file.

This ensures that a real difference is an error, instead of a warning,
and that rolling a new version of the compiler or plugin is not going
to break things needlessly (which was the reason for the warning in
the first place).

Bug: 72810
Change-Id: Icad40dee8d70bf8882e5c6c2b18b6268717f7e8f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/561001
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>