[paver][flake] Wait for ramctl before using it
There was a race here where we are attempting to use a device before it
has finished enumerating. Prior to moving ramdisk, this race was always
won, but now that it's downstream of platform-bus, it sometimes loses
the race.

Fixed: 81517
Change-Id: I04f8530090ec16b349884016283d24d64e646021
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/560966
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>