[sshutil] Consider EOF to be a ConnectionError
An EOF when trying to run an SSH command likely indicates that the SSH
client was closed due to a failed keep-alive ping. This should be
classified as a connection error so that testrunner will retry the
command instead of classifying it as a test failure in case an SSH flake
manifests in this way.

Fixed: 81609
Change-Id: I2725f5ab250227bef32af7e9c0824d57ab3e297d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/561721
Reviewed-by: Ina Huh <ihuh@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>