[kernel][vm] Eliminate function scoped statics from arm platform
This change is part of a series that eliminates from the kernel
function scoped static variables that require dynamic initialization.
In the kernel environment function scoped statics cannot be both
safely and efficiently initialized so we are replacing them with
alternatives.

In platform/generic-arm/platform.cc, hoist panic_started and halted to
files scope.

Bug: 79612
Change-Id: Ib36838ea716d3d92b4a896343bb0e5a2cc06f476
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/560641
Reviewed-by: John Grossman <johngro@google.com>
Commit-Queue: Nick Maniscalco <maniscalco@google.com>