[kernel]: Remove serial acpi handlers.
As of recently, the boot shims parse the ACPI table descriptors and
synthesize a serial zbi item, which contains the configurations from the
uart in the acpi table.

Also as of recently the command line is pulled from BootOptions that
already handles this cases, in which it ends up using the zbi item for
its serial configuration

This CL just removes dead code, as a result of the mentioned changes.

Bug: 80903 # Fixed
Test: CQ
Change-Id: Ie227a740a291f9dd71a0518420b1fe272578a470
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/561422
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>