[fidl][summary] Start version tracking summaries
This commit adds the API summary files [1] to each FIDL library.
The idea is that from a certain point onwards, we keep these files in
sync with the contents of the library.

The presence of these files is not yet enforced.  Enforcement will be
turned on once all FIDL files which need this treatment have been
equipped with the summary files.  Enforcement is similar to now `.api`
files are handled today.

[1]: https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0076_fidl_api_summaries?hl=en

Bug: 68951
Change-Id: I0fede3fb9abb4738b4795f9160d4102055aa0fd5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/537862
Reviewed-by: Alice Neels <neelsa@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Felix Zhu <fcz@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>