[wlan][sme] Use ServingApInfo in ClientStatusResponse
There are several related refactors in this change designed to improve
the readability of types in wlan-sme. The primary motivating refactor
is the creation of fidl_sme::ServingApInfo to decouple
the data structure for connection status from the data structure for a
scan result. Both were previously represented by fidl_sme::BssInfo
which is not well-suited for both use cases.

* Created fidl_sme::ServingApInfo

The term ServingApInfo comes from "serving access point" in IEEE
802.11-2016 Section 3.1 where it's defined as "The AP to which the
station (STA) is associated." This is more precisely what the data
structure describes in a ClientStatusResponse.

* fidl_sme::ClientStatusResponse is now an enum to remove ambiguity

The previous definition was a struct with an optional `connected_to`
field and a required `connecting_to_ssid` which was sometimes filled
with an empty vector. The new type is much more explicit and does not
implicitly requiring filling fields with null-like types to represent
unnamed states.

See src/connectivity/wlan/lib/sme/src/client/inspect.rs:175 for an
example of this problem in action.

* Deleted convert_bss_description_skip_compatibility_check()

This function was obviously a hack and is now removed.

* Renamed wlan_sme::client::Status to wlan_sme::client::ClientSmeStatus

This makes room for a future AP status type and other status types not
related to the SME itself. It's also not unreasonable to expect all
status types to be moved up to a separate `wlan_sme::status` in the
future.

Test: fx test //src/connectivity/wlan
Bug: 79613

Change-Id: I73e37a263ef39bdb53bd25d9dbdd7f10a8baeec6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/554125
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>
Reviewed-by: Rebecca Silberstein <silberst@google.com>
Reviewed-by: Kiet Tran <kiettran@google.com>
API-Review: Rebecca Silberstein <silberst@google.com>