[kernel] Remove spurious error logging in zx_system_mexec_get_payload()
zx_system_mexec_payload_get() is a bit funny in that it writes to a
user-provided buffer, but the user has no idea how large to make that
buffer. The workaround to that is that the syscall will return
ZX_ERR_BUFFER_TOO_SMALL in that case, and the user would react by making
the syscall again with a larger buffer. This change removes error
printing in the branch that returns ZX_ERR_BUFFER_TOO_SMALL; it's
spurious given the above.

Change-Id: I8cdb4589cbe1929c6eed43cc12d34c64f3f2b40a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/561507
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>