[blobfs] Move merkel mapping to blob verifier.
Before, the Blob kept the fzl::OwnedVmoMapper associated with the Merkle
tree. A pointer to this was given to the BlobVerifier, but this is
stored very far away such that the relationship of the ownership is not
clear. I believe during cleanup the verifier actually has an invalid
pointer, but we're OK because we don't try to verify anything at this
point.

This patch moves ownership of the mapping to the BlobVerifier so it can
be closer to the pointer that references it. By passing the BlobLayout
in to the verifier, the API could also be simplified in various ways.

Bug: 79611
Change-Id: If561a3880cf1c5ff8c87a39855ef699e62ac5e1c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/561890
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Stephen Demos <sdemos@google.com>