[fidl][transformer] More V2 GIDL tests & fix transformer bug
This CL adds a variety of V2 GIDL tests to the existing ones.
There are still a number that still need to be converted.

When enabling these tests, a bug in the transformer was
discovered. Namely that it in some cases it was possible to
read and write past the buffer size.

Change-Id: Ie0fbf91d6b7bb5391d7e0193044a6ee84090892e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/561662
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>