[power] Use new connect_to_driver implementation
Replaces the crate::utils::connect_to_driver function with a similar
implementation from elsewhere in the codebase. The new implementation
was recommended by the FDF team while investigating a bug in connecting
to a driver's topological path. The bug would result in the Power
Manager attempting to connect to a driver before it was fully ready to
receive requests (likely related to improper directory file watching).

The new implementation also has the benefit of using async Rust crates
for file watching. This is a clear improvement and is preferred over the
existing implementation's usage of blocking FDIO calls.

Test: fx test power-manager-tests
Change-Id: Ia991ed6b1fc8e55eb77ec8f2af46319f52994f74
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/558750
Commit-Queue: Patrick Shickel <pshickel@google.com>
Reviewed-by: Onath Claridge <claridge@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>