[build][go] Hardlink instead of symlinking source files
This makes it possible to use `//go:embed` directives in Go code as long
as the embedded file is declared in the `sources` of the GN target that
builds the Go file. `//go:embed` doesn't support symlinks.

This is a bit of a hacky workaround; a better long-term solution is
described in the linked bug.

Bug: 81748
Change-Id: I92ed082ec4f564894ed5850c58bf2da7d1ed1a47
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/562221
Commit-Queue: Oliver Newman <olivernewman@google.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>