[blobfs] Remove BlobLoader::LoadResult.
Remove the BlobLoader::LoadResult and put all information in the
LoaderInfo struct.

The main structure change is that a BlobLayout object is now kept for
every blob. Before the BlobLayout was used to compute a few values which
were saved, and then it was discarded. The BlobLayout struct is
relatively small, this allows less duplication, and it seems like the
values it has should be readily available anyway. This allows some
cleanup in various places.

Bug: 79611
Change-Id: I558df838acbe0113355470b456a5e15f3b0d91a1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/562143
Reviewed-by: Stephen Demos <sdemos@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>