[ermine] Handle invalid focus handle from focus chain
Due to a recent change: https://fxrev.dev/561821 (now reverted) the
focus service encountered invalid focus chain. The shell should always
be in the focus chain. But it was receiving a focus chain with only the
root view in it. This results in an invalid app state and an
exception because settings start was called consecutively.

This change adds a guard to only respond when the focus is changing,
basically handle invalid focus chains.

Change-Id: I5d7501f209cada155baebfaa0ee63a841a4c9e9c
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/562644
Fuchsia-Auto-Submit: Sanjay Chouksey <sanjayc@google.com>
Reviewed-by: Yeonhee Lee <yhlee@google.com>
Commit-Queue: Sanjay Chouksey <sanjayc@google.com>