[ffx][test] Remove experimental flag from ffx test
ffx test is intended to replace ffx component test. While ffx test is
not completely ready, the parts of it that replace the functionality of
ffx component test are ready. This does a few things with the goal of
encouraging users to migrate to ffx test while marking the areas which
are still experimental:
* Add a deprecation notice to ffx component test pointing at ffx test
* Remove the experimental flag on ffx test
* Add an experimental flag around features of ffx test that rely on
the unstable structured test output

Change-Id: I25a8ff9bdfdf0e4eb880f0318160a8b95de195c5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/550828
Commit-Queue: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Matthew Boetger <boetger@google.com>