[zircon] HACK: Replace the BrwLock in the handle table with a Mutex
The BrwLock implementation can not work on riscv, it uses a 128 bit "atomic"
compare exchange which doesn't have an instruction on riscv. To progress
on the port, this replaces all BrwLock usages with a Mutex and remove
the guard the offending atomic operation with a __riscv macro check to
let LLVM compile the kernel.

We can't reasonably keep this long-term because it butchers performances
but well, it boots. ¯\_(ツ)_/¯

Change-Id: Iaa08f3cfd034f3fe92bc92526ea44db881345d8b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553624
Reviewed-by: Roger Tawa <rogerta@google.com>