[blobfs] Make reserving blocks more efficient
Skip past all allocated blocks in a region in FindUnallocatedExtent
when trying to find free blocks. Previously if we were trying to
reserve 1 block and the first free block was the 14000th (120MB used
with no holes) then it would take 14000 calls to FindUnallocatedExtent
and 28000 calls to RawBitmap::Scan to find it. This was because
FindUnallocatedExtent would only advance the number of blocks we were
looking for on each call. Now FindUnallocatedExtent advances entire
allocated regions in a single call.

Change-Id: I13d6753a21fd530628b0616b44eac35aeec7967e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/562161
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Chris Drouillard <cdrllrd@google.com>