[component_manager] don't fail to destroy an instance with bad storage
Failing to route a storage capability would cause instance destruction
to fail in component manager, because it was trying to delete the
instance's storage. If an instance has an invalid route for a storage
capability, then it was never able to write to storage in the first
place, and component manager can safely ignore the error.

Fixed: 81289
Change-Id: Ie4b179761a5fc43c6be332bbb4875f6cf4ee002d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/559201
Commit-Queue: Derek Gonyeo <dgonyeo@google.com>
Reviewed-by: Gary Bressler <geb@google.com>