[debugger] Enhance cross-platform errors.
Adds some non-system-derived error values to the cross-platform
debug::Status struct for the types of errors we expect the frontend
needs to handle specially. This also gives a direct mapping to the
frontend's Err value.

This brings debug::Status and zxdb::Err closer, we may want to consider
merging these in the future.

Currently these values are just derived from the zx_status_t values at
the time of creation so there should be no effective change. But by
removing checks for specific zx_status_t values in the frontend we have
more flexibility in being precise in the backend implementation and not
relying on potentially changing parts of the system.

As part of testing, I noticed that all "run" commands hang on Fuchsia so
some better error messaging is added in this case.

Change-Id: I53ca4e9b3e9acfc686a764e02ba87da1d71f9854
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/562812
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Dangyi Liu <dangyi@google.com>