[build][sysmgr] Allow duplicate sysmgr config files.
It is illegal to specify a build configuration where multiple
sysmgr config files define the provider of a single service.
However, it's not hard to inadvertently create a situation
where the *identical* config file appears twice.  Instead of
dying with a confusing build error, this CL detects and allows
this situation.

Change-Id: Iba5b07b5a2ac957cc0f0c27c847f8a9c38ad3439
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/554969
Reviewed-by: Bryan Henry <bryanhenry@google.com>
Commit-Queue: Josh Gargus <jjosh@google.com>