[virtualization][tests] Split hypervisor_tests into per-arch files.
hypervisor_test currently consists of a single C++ file and two
architecture-specific assembly files. The current set of tests
contain a mix of architecture-dependent and -independent tests
and helpers.

This CL moves the hypervisor tests into its own subdirectory, and
splits the main test file into three, with a layout as follows:

  hypervisor/
    hypervisor_tests.h  (new header to allow use of helpers)
    constants.h         (previously "hypervisor_tests_constants.h")
    main.cc             (common helpers and main function from
                        "hypervisor_test.cc")
    common_tests.cc     (common tests from "hypervisor_test.cc")
    arch/
      arm64/
        guest_code.S    (previously "hypervisor_tests_arm64.S")
        arch_tests.cc   (ARM64-specific parts of "hypervisor_tests.cc")
      x64/
        guest_code.S    (previously "hypervisor_tests_x64.S")
        arch_tests.cc   (x64-specific parts of "hypervisor_tests.cc")

This reduces the amount of conditionally-compiled code, and makes it
clearer to readers which tests apply to which architecture.

The split isn't perfect: some conditional compilation is still present
in the common headers and code. Future CLs will address this.

Change-Id: I4b3d6c14fa3e1c805d5b6af3b26a3aae1877e102
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/559905
Commit-Queue: David Greenaway <dgreenaway@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>