[scrutiny] Improve logging for verify_routes.
Since verify_routes is a complied_action that is part of the build it
makes more sense for critical failures to present themselves in the
terminal rather than in /tmp/scrutiny.log. This change adds a simple
logger implementation for critical errors and adds some basic context to
the component_tree.

Bug: 81806
Change-Id: I77476dd4e2ad810c701cf171bcb5321feb219627
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564081
Reviewed-by: Drew Fisher <zarvox@google.com>
Commit-Queue: Benjamin Wright <benwright@google.com>