[scrutiny] Improve collection error logging.
Prior to this change failures in retrieving data collectors would return
a uuid that was very difficult to decode. This change provides a real
name to each collector and a short description of what is in the
collection. This greatly improves the ability to debug when a collection
cannot be processed.

Bug: 81806
Change-Id: Ib54faab4e300e5bcad85831b4a66b78bafd4e6fd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564121
Commit-Queue: Benjamin Wright <benwright@google.com>
Reviewed-by: Drew Fisher <zarvox@google.com>