[fs] Improve filesystem ID handling.
The Factoryfs filesystem ID was uninitialized. Hook this up.

Compute the blobfs legacy filesystem ID on demand. This is relatively
efficient and uncommon so doesn't warrant caching on the Blobfs object.

Store the zx::event in Minfs and Memfs. Previously these discarded the
event which negated the guarantee that the id is globally unique.

Some errors around creating and duplicating these events are now
ignored. The only time these will fail for non-misuse is if the kernel
is out of memory and in this case, prserving the exact error is unlikely
to be useful.

Change-Id: Idc54e3f960f6e2a4cb5e0089f6af7c3a828e36d4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564123
Reviewed-by: Stephen Demos <sdemos@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>