[assembly] Add options to modify update/base package names
Enable clients of assemble_system to specify the name of their update
and base packages. This is useful specifically for the "prime" images,
where we generate separate base and update packages and publish them to
the same repository as older base and update packages, but we do not
want the new packages to override the old.

Bug: 79716
Change-Id: Ife2f5d5ac70bf6dee47399854bb19df2ee81a9ff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/563168
Commit-Queue: Aidan Wolter <awolter@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>