[wlan] Respect scan type in `wlan-dev connect`.
This change respects the `--scan-type` option in the `wlan-dev` tool's
`connect` command. After the removal of join scans in the SME,
`wlan-dev` temporarily performs these scans itself before attempting to
connect, but always performed an active scan regardless of the requested
scan type.

Today, this means that `wlan-dev connect` commands will not work if the
active scan type is not supported by the driver, as is currently the
case in `iwlwifi`.

Change-Id: Ife0724f69f97ae6be40f2f7517566a73e6b5d68c
Test: Manual testing and `wlantool-tests`.
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/562145
Commit-Queue: Sean Olson <seanolson@google.com>
Reviewed-by: Kiet Tran <kiettran@google.com>
Reviewed-by: Charles Celerier <chcl@google.com>