[scrutiny] Toolkit takes no dependency on the model.
The original intention of the toolkit plugin was to be a standalone thin
wrapper around Scrutiny utility code that allows the extraction of
blobfs and fvm. With the introduction of the packages toolkit option
that provides support for package extraction by urls this was no longer
possible. This change moves the package extractor into core and removes
the dependency from toolkit. Since toolkit is currently load bearing in
the verify_build python script this also prevents the Data Collector
being invoked in that script which is blocking landing early scheduler
failures.

Bug: 81806
Change-Id: I8a1945e3f959b4bb95443b76629f861e03d70c27
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564134
Reviewed-by: Andrew Davies <awdavies@google.com>
Reviewed-by: Drew Fisher <zarvox@google.com>
Commit-Queue: Benjamin Wright <benwright@google.com>