[ffx][test] Improve message on error
Instead of returning a normal error, use ffx_bail when a test
fails.
When an error is returned noramlly, ffx prepends a 'BUG:' prefix. The
prefix is not added when ffx_bail is used. In most cases, when a test
fails this is not a bug in ffx or the test framework.

Testing: manual verification with `fx test` against a failing test.

Change-Id: I8e86954951738c6e79d639bea8a8cb3ae8e017f3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/563391
Commit-Queue: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>