[ffx] Improve speed of `ffx component run-legacy`
This plugin was using old `cs` code to traverse the hub and read
process IDs so that it could print them out when the component
was started.

Unfortunately, this functionality was not working properly because
the hub was read before the component was started.

Removing this functionality increases the speed of this command
dramatically.

As a result, `fx mem` can now use `ffx component run-legacy` again.

Fixed: 81144
Test: fx test ffx_component_run_legacy
Change-Id: I9702d8fff0eece0ac5f9d981dc1c2e79c92e04cf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564005
Commit-Queue: Xyan Bhatnagar <xbhatnag@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Jordon Wing <jwing@google.com>