[build] Build packages.tar.gz from correct inputs.
Changes the way the packages.tar.gz archive is built.
Before this CL, the 'tarmaker' tool was called to grab
whatever was under $BUILD_DIR/amber-files/{keys,repository}/.

With this CL, the content of the archive is entirely
determined from all_packages_manifest.list and the targets.json
file produced by //build/images:publish, which lists which
packages and data should be available in the archive.

For testing, the content of the archive is compared
to out/default/amber-files/{keys,repository} from a
clean (non-incremental build) to verify that they are
identical! E.g.:

  fx ninja -C out/default -t clean
  fx ninja -C out/default packages.tar.gz
  mkdir /tmp/aaa && cat out/default/packages.tar.gz | tar -C /tmp/aaa -zxf-
  diff -qurN out/default/amber-files /tmp/aaa/amber-files \
    && echo "OK :-)" || echo "KO :-("

Bug: 81322
Change-Id: Iaf0be4e983f87edd3f471661ccc66d5f5d0fbe54
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/559203
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: David Turner <digit@google.com>