[bt][rfcomm] Increase buffer size for incoming/outgoing frames
Increase the buffer capacity for frames received by the peer and frames
to be sent to the peer. In some cases, there may be multiple frames
received at once (before being processed). For example, when a profile
client (like HFP) wants to send data to the peer for the first time, the
RFCOMM component first sends a ModemStatus command to indicate
readiness followed by the user data. Increasing the buffer size allows
the implementation to maintain a reasonably sized queue of messages to
be sent/received to/from the peer.

Add an integration test that verifies that the initiating RFCOMM
component can successfully establish an outgoing RFCOMM channel to the
responding RFCOMM component. Validate that user data can be exchanged
between the two channels - this check uncovered the subtle bug w.r.t the
buffer capacity in which the first user data frame was not being sent
due to a lack of space.

Bug: 77372
Test: fx test bt-rfcomm-integration-tests
Multiply: bt-rfcomm-integration-tests

Change-Id: I2d8975e9838a5ddc838152254b2078283570dd1d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/562262
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Marie Janssen 💖 <jamuraa@google.com>