[weave] Rebind lowpan device on device operation
If the lowpan service restarts or otherwise closes the connection to
weavestack, the connection must be re-established before any further
provisioning or status operations. Before these operations, rebind to
the lowpan service, giving us an opportunity to establish the connection
or fail if the lowpan service is down.

Bug: b/195447349
Test: fx test weavestack-unittests
Test: fx test weavestack-adaptation-unittests
Change-Id: I6711025fca1abfd60790bf3cd52735c57c48ecce
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564341
Reviewed-by: Jason Graffius <jgraff@google.com>
Commit-Queue: Prashanth Swaminathan <prashanthsw@google.com>