[fxfs] Fix tombstone issue
Tombstone records didn't play nicely with directory records.  After
inserting a tombstone record, it's possible that directory child records
might exist in other layers, but the lookup function doesn't merge with
tombstone records, so it was possible to perform a lookup on a deleted
directory and find an old child.

To address this, we use the same key as object records but with a value
of None so that we will always know a deleted directory when we try to
open it.

In addition, we keep track of whether a directory is deleted within the
object-store layer, and return appropriate errors for all operations.

This addresses a flake which would manifest itself as an object
reference mismatch, since it could lead to updates on a deleted
directory which would resurrect the directory.

Change-Id: Icb5fc05cf940816df8d678df02fd0dcc512b5fb3
Multiply: fxfs-tests: 100
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564361
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>