[omaha-client] Policy return minimum wait for standard logic.
When using standard logic, even though we have accurate monotonic time,
it is possible that updates has been throttled, if a scheduled update
gets throttled, it will come back to ask `compute_next_update_time`
again without updating `last_update_time`, so in this case it could
return a time in the past or future, depends on the fuzzing, but
eventually it will always return a time in the past and results in an
infinite loop until the rate limit period is passed.

This CL fixes this case by adding minimum wait to the returned
`CheckTiming` even in standard logic.

Fixed: 82023
Change-Id: I41a1a6b180de586d99c3c7b7deddd73a7a0c2fd6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/563926
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Sen Jiang <senj@google.com>