[fidl] Add anonymous layout usages to goldens
A new golden file containing anonymous layouts is added in
preparation for any special support that is added in the bindings
(i.e. so that behavior changes are observable in the goldens).

The feature gating of anonymous layouts is also removed to make
testing easier; This enables anonymous layouts before
fxbug.dev/74683 is fully finished but that's OK because
1. The remaining CLs (stacked above) will land shortly, and
2. The remaining changes won't break any users that decide
   to use the feature before 1

Test: fx regen-goldens
Change-Id: Idc6657b45f6ecf0ae78843f34e1f348dc681612b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/554981
Commit-Queue: Felix Zhu <fcz@google.com>
Reviewed-by: Alex Zaslavsky <azaslavsky@google.com>