[network/tests] Log link local addresses on fail
Before this change, we see a rare flake that produces the error:

  Expected equality of these values:
    got_unknown_link_local_addrs
      Which is: 3
    want_unknown_link_local_addrs
      Which is: 4

Which is not actionable. Instead of counting the addresses, collect them
into a vector and assert on its size; on failure, the matcher will emit
the contents of the vector, producing a more actionable message.

Bug: 82098
Change-Id: I7df1669492792217a1184e26233788ed006f5ae7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/565123
Fuchsia-Auto-Submit: Tamir Duberstein <tamird@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>