[scrutiny_config] Extract config into its own crate.
The scrutiny configuration crate should be independent of deps on
the framework. This allows for the flexibility of both the frontend
and the framework depending on the generic configuration. This is part
of a larger refactoring to improve build time configuration of Scrutiny
inside compiled actions.

Bug: 81808
Change-Id: I911f9e38c9579933bc39a047ef4e21f5a684aa14
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564669
Reviewed-by: Drew Fisher <zarvox@google.com>
Reviewed-by: Mark Dittmer <markdittmer@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Benjamin Wright <benwright@google.com>