[fidl][transformer] Fix bugs found by fuzzing the transformer
I manually ran the fuzzer on LLCPP + the transformer and it hit a small
number of assertions. This CL provides fixes so that it will return
cleanly and not hit an assertion at runtime.

In the future, I hope to run the fuzzer automatically on v2.

Bug: 79177
Change-Id: I26edb715f67964dbc7f064b30000d5ea90b56d26
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/565163
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>