[sys][testing] TestWithEnvironmentFixture rename for soft migration
`TestWithEnvironment` will migrate to the SDK, without gTest dependencies.

This CL adds `TestWithEnvironmentFixture`, to replace `TestWithEnvironment`
usage, in- and out-of-tree, so existing out-of-tree dependencies on
`TestWithEnvironment` can be updated, first to `TestWithEnvironmentFixture`,
and later to the revised `TestWithEnvironment`.

This is the first migration step to resolving fxbug.dev/81468.

Change-Id: I095810f8dff5977479faf90654333b5479592a32
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564570
Reviewed-by: David Worsham <dworsham@google.com>
Reviewed-by: Gabe Schine <thatguy@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Rich Kadel <richkadel@google.com>