[ffx][daemon] Move target et al into standalone lib.
This is the first (simplest) step for moving/refactoring target.rs as
well as making `Target` and `TargetCollection` visible in the services
library without making a circular dependency of daemon <-> service_lib.

This moves the following into its own crate:
* target.rs
* fastboot mod
* logger mod
* onet.rs
* ssh.rs
* zedboot.rs

The only notable changes are that many `pub(crate)` declarations were
changed to `pub` and `#[cfg(test)]` was removed from functions and
replaced with a doc comment stating the function is "intended for
testing." This was all done in lieu of a larger breakdown of
files/refactoring.

This also does a long-needed deps cleanup of the BUILD.gn files for target
and for the daemon.

Bug: 81488
Change-Id: I2a85a704ee061cfa5bc1770d479409922f59db08
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/562223
Fuchsia-Auto-Submit: Andrew Davies <awdavies@google.com>
Commit-Queue: Andrew Davies <awdavies@google.com>
Reviewed-by: Andy Weiss <dragonbear@google.com>