[fidl][transformer] Add more assertions about size bounds
This CL isn't in response to a particular issue - it was noticed that
primitive / enum / bits / handle / struct perform memcpy without a
immediately previous call to IncreaseNextOutOfLine. This means the
correctness of previous out of line block calculations is assumed and
the assertions added in this CL provide additional protection.

In addition, all assertions were changed to debug assertions because
that was the original intent.

Bug: 79177
Change-Id: I11eeccfc37ddb63aa7ac5d5d22d467f2aabb3398
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/565121
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>