[scrutiny] Remove ModelEnvironment in favor of ModelConfig.
Prior to this change the ModelEnvironment basically just copied all the values
from the ModelConfig. This was because the scrutiny_frontend crate dependend
on scrutiny_framework. Now scrutiny_config breaks this cycle so its possible
for scrutiny_framework to dependend on and use scrutiny_config. This
makes it simpler to add configuration for all collectors to use. This is
important in making Scrutiny more flexible in accepting variables from
the build in compiled actions.

Bug: 81088
Change-Id: Iaee5dc98d6bbd52d1990a668b22a9f00fdfe769a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564672
Commit-Queue: Benjamin Wright <benwright@google.com>
Reviewed-by: Drew Fisher <zarvox@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>