[memfs] Add dispatcher to memfs.
In the future, the ManagedVfs will require a dispatcher. Convert memfs
to always require a dispatcher in the constructor.

Initialize some class members in the header for clarity.

Make GetMinfsBlksize to GetPageSize() and de-inline it (an inline
function with a static is suspicious).

Remove the name from the Memfs constructor (it was never used) and
convert the name in ::Create to a string_view (matches its eventual use
in a Dnode).

Change-Id: I9a187e8a32c42d4b935a6e779247ddabf992f4e0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564135
Reviewed-by: Brandon Castellano <bcastell@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>