[ffx][inspect] Fix hang from empty results
Requesting a snapshot from an ArchiveReader can return empty results if
for example you are asking for data for a non-existent component. The
iquery code explicitly sets this flag to not retry on an empty result. I
forgot to carry that over to the ffx plugin code that interacts with the
ArchiveReader.

Fixed: 81981

Change-Id: I351255cd92aba2ba0cb883b9b895381abc4395ff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564322
Reviewed-by: Amit Uttamchandani <amituttam@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Andy Weiss <dragonbear@google.com>