[gigaboot] remove network spam
We had a few log messages that were printed whenever we saw a UDP packet
not meant for us, but this is a normal occurrence, commonly via unicast
flooding when a switch doesn't know which port a device is connected to
so has to send the packet to all of them.

Remove the logs, they're not helpful and get spammy if the user has a
network that triggers them often.

Bug: 81988
Change-Id: Iadc661c48b15be6783a0bc8483b5ac04fceb355d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/565885
Reviewed-by: Varun Sharma <vars@google.com>
Commit-Queue: David Pursell <dpursell@google.com>