[sdk] Add blobfs to the partner SDK
The image assembly tool (RFC-0072) requires the blobfs tool to construct
a BlobFS image. We are adding blobfs to the SDK so that the image
assembly tool is usable out of the tree.

We are naming the tool blobfs_do_not_depend to make it obvious that the
CLI surface is not expected to be used directly by anything outside of
the SDK, and is an implementation detail of the Image Assembler.

Change-Id: I9269f1b94da5baedecad0011b5a4f28a1e29bc1d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/541143
Reviewed-by: James Sullivan <jfsulliv@google.com>
Reviewed-by: Brett Wilson <brettw@google.com>
Reviewed-by: Misha Gridnev <gridman@google.com>
API-Review: Jeremy Manson <jeremymanson@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>